feat: store session in zustand
