+++ src/App.tsx
@@ -5,8 +5,27 @@ import NavBar from "./components/NavBar";
 import Login from "./pages/Login";
 import Footer from "./components/Footer";
 import { Account } from "./pages/Account";
+import { useEffect } from "react";
+import { supabase } from "./utils/supabaseClient";
+import { useAuthStore } from "./store/useAuthStore";
 
 const NavbarWrapper = () => {
+  const updateSession = useAuthStore((state) => state.updateSession);
+
+  useEffect(() => {
+    supabase.auth.getSession().then(({ data: { session } }) => {
+      updateSession(session);
+    });
+
+    const {
+      data: { subscription },
+    } = supabase.auth.onAuthStateChange((_event, session) => {
+      updateSession(session);
+    });
+
+    return () => subscription.unsubscribe();
+  }, []);
+
   return (
     <div className="grid min-h-screen w-full grid-rows-layout items-center gap-8 bg-gradient-to-b from-neutral-800 to-neutral-900 py-8 px-12 text-white">
       <NavBar />
